rename context file to McAppContext
