+++ src/components/AsyncCard.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { AsyncStatus } from '../contexts/CommAppContext';
+import { AsyncStatus } from '../contexts/McAppContext';
 import SkeletonLoadingPage from './loading/SkeletonLoader';
 import EdsCard from './EdsCard';
