+++ src/components/AsyncPage.tsx
@@ -1,6 +1,6 @@
 import { Banner } from '@equinor/eds-core-react';
 import React from 'react';
-import { AsyncStatus } from '../contexts/CommAppContext';
+import { AsyncStatus } from '../contexts/McAppContext';
 import { COLORS } from '../style/GlobalStyles';
 import EdsIcon from './icons/EdsIcon';
 import SkeletonLoadingPage from './loading/SkeletonLoader';
